+++ src/model/Layers/Hidden_Layer.cpp
@@ -1,8 +1,9 @@
 #include "Hidden_Layer.hpp"
+#include <cstddef>
 
 namespace nn {
 void Hidden_Layer::forward(const std::vector<Global::ValueType> &metrix) {
-	for (int i = 0; i < dots.size(); i++) {
+	for (size_t i = 0; i < dots.size(); i++) {
 		dots.net[i] = 0;
 		for (size_t j = 0; j < metrix.size(); j++) {
 			dots.net[i] += Parameters.weights[i][j] * metrix[j];
