+++ src/model/model.hpp
@@ -23,10 +23,10 @@ class model {
 	void reset();
 	Layer &getLayer(const int i) { return *network.layers.at(i); }
 	void updateWeights(const gradient &gradients);
-	int getOutputSize() const { return network.config.output_size; }
-	int getInputSize() const { return network.config.input_size; }
-	int getHiddenLayerCount() const { return network.config.hidden_layer_count(); }
-	int getLayerCount() const { return network.config.hidden_layer_count() + 1; }
+	size_t getOutputSize() const { return network.config.output_size; }
+	size_t getInputSize() const { return network.config.input_size; }
+	size_t getHiddenLayerCount() const { return network.config.hidden_layer_count(); }
+	size_t getLayerCount() const { return network.config.hidden_layer_count() + 1; }
 	const bool useVisual;
 };
 } // namespace nn
