+++ src/model/neuralNetwork.cpp
@@ -10,7 +10,7 @@ neural_network::neural_network(const NetworkConfig &network_config)
 	layers.reserve(network_config.hidden_layer_count() + 1);
 
 	int prev_size = network_config.input_size;
-	for (int i = 0; i < network_config.hidden_layer_count(); i++) {
+	for (size_t i = 0; i < network_config.hidden_layer_count(); i++) {
 		layers.emplace_back(std::make_unique<Hidden_Layer>(
 		    network_config.layers_config[i].size,
 		    prev_size,
